Implement review feedback: Place Wasm DSL and associated instruction classes in a separate file and namespace
