+++ src/coreclr/tools/Common/Compiler/ObjectWriter/WasmNative.cs
@@ -5,6 +5,7 @@ using System;
 using System.Diagnostics;
 using System.Linq;
 using System.Collections.Generic;
+using ILCompiler.ObjectWriter.WasmInstructions;
 
 namespace ILCompiler.ObjectWriter
 {
@@ -192,163 +193,6 @@ namespace ILCompiler.ObjectWriter
                 return $"(func (param {paramList}))";
             return $"(func (param {paramList}) (result {returnList}))";
         }
-    }
-
-    // Represents a WebAssembly expression used in simple contexts for address calculation
-    public enum WasmExprKind
-    {
-        I32Const = 0x41,
-        I64Const = 0x42,
-        GlobalGet = 0x23,
-        I32Add = 0x6A,
-    }
-
-    public static class WasmExprKindExtensions
-    {
-        public static bool IsConstExpr(this WasmExprKind kind)
-        {
-            return kind == WasmExprKind.I32Const || kind == WasmExprKind.I64Const;
-        }
-
-        public static bool IsBinaryExpr(this WasmExprKind kind)
-        {
-            return kind == WasmExprKind.I32Add;
-        }
-
-        public static bool IsGlobalVarExpr(this WasmExprKind kind)
-        {
-            return kind == WasmExprKind.GlobalGet;
-        }
-    }
-
-    class WasmInstructionGroup : IWasmEncodable
-    {
-        readonly WasmExpr[] _wasmExprs;
-        public WasmInstructionGroup(WasmExpr[] wasmExprs)
-        {
-            _wasmExprs = wasmExprs;
-        }
-
-        public int Encode(Span<byte> buffer)
-        {
-            int pos = 0;
-            foreach (var expr in _wasmExprs)
-            {
-                pos += expr.Encode(buffer.Slice(pos));
-            }
-            buffer[pos++] = 0x0B; // end opcode
-            return pos;
-        }
-
-        public int EncodeSize()
-        {
-            int size = 0;
-            foreach (var expr in _wasmExprs)
-            {
-                size += expr.EncodeSize();
-            }
-            // plus one for the end opcode
-            return size + 1;
-        }
-    }
-
-    public abstract class WasmExpr : IWasmEncodable
-    {
-        WasmExprKind _kind;
-        public WasmExpr(WasmExprKind kind)
-        {
-            _kind = kind;
-        }
-
-        public virtual int EncodeSize() => 1;
-        public virtual int Encode(Span<byte> buffer)
-        {
-            buffer[0] = (byte)_kind;
-            return 1;
-        }
-    }
-
-    class WasmConstExpr : WasmExpr
-    {
-        long ConstValue;
-
-        public WasmConstExpr(WasmExprKind kind, long value) : base(kind)
-        {
-            if (kind == WasmExprKind.I32Const)
-            {
-                ArgumentOutOfRangeException.ThrowIfGreaterThan(value, int.MaxValue);
-                ArgumentOutOfRangeException.ThrowIfLessThan(value, int.MinValue);
-            }
-
-            ConstValue = value;
-        }
-
-        public override int EncodeSize()
-        {
-            uint valSize = DwarfHelper.SizeOfSLEB128(ConstValue);
-            return base.EncodeSize() + (int)valSize;
-        }
-
-        public override int Encode(Span<byte> buffer)
-        {
-            int pos = base.Encode(buffer);
-            pos += DwarfHelper.WriteSLEB128(buffer.Slice(pos), ConstValue);
-
-            return pos;
-        }
-    }
-
-    // Simple DSL wrapper for creating Wasm expressions
-    static class Global
-    {
-        public static WasmExpr Get(int index)
-        {
-            return new WasmGlobalVarExpr(WasmExprKind.GlobalGet, index);
-        }
-    }
-
-    static class I32
-    {
-        public static WasmExpr Const(long value)
-        {
-            return new WasmConstExpr(WasmExprKind.I32Const, value);
-        }
-
-        public static WasmExpr Add => new WasmBinaryExpr(WasmExprKind.I32Add);
-    }
-
-    class WasmGlobalVarExpr : WasmExpr
-    {
-        public readonly int GlobalIndex;
-        public WasmGlobalVarExpr(WasmExprKind kind, int globalIndex) : base(kind)
-        {
-            Debug.Assert(globalIndex >= 0);
-            Debug.Assert(kind.IsGlobalVarExpr());
-            GlobalIndex = globalIndex;
-        }
-
-        public override int Encode(Span<byte> buffer)
-        {
-            int pos = base.Encode(buffer);
-            pos += DwarfHelper.WriteULEB128(buffer.Slice(pos), (uint)GlobalIndex);
-            return pos;
-        }
-
-        public override int EncodeSize()
-        {
-            return base.EncodeSize() + (int)DwarfHelper.SizeOfULEB128((uint)GlobalIndex);
-        }
-    }
-
-    // Represents a binary expression (e.g., i32.add)
-    class WasmBinaryExpr : WasmExpr
-    {
-        public WasmBinaryExpr(WasmExprKind kind) : base(kind)
-        {
-            Debug.Assert(kind.IsBinaryExpr());
-        }
-
-        // base class defaults are sufficient as the base class encodes just the opcode
     }
 
     public abstract class WasmImportType : IWasmEncodable
